+++ tracker_project/tracker/models.py
@@ -5,7 +5,7 @@ class BMI(models.Model):
     age = models.PositiveIntegerField(blank = True, null = False)
     heightFeet = models.FloatField(blank = True, null = False)
     heightInches = models.FloatField(blank = True, null=True)
-    heightMetres = models.PositiveIntegerField(blank = True, null=False)
+    heightMetres = models.FloatField(blank = True, null=False)
     weightPounds = models.FloatField(blank = True, null=True)
     weightKilos = models.FloatField(blank = True, null=True)
